+++ callisto_core/delivery/management/commands/find_matches.py
@@ -1,11 +1,11 @@
 from django.core.management.base import BaseCommand
 
-from callisto.delivery.matching import MatchingApi
+from ....utils.api import MatchingApi
 
 
 class Command(BaseCommand):
     help = 'finds matches and sends match reports'
 
     def handle(self, *args, **options):
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.stdout.write('Matching run')