+++ callisto_core/notification/models.py
@@ -6,8 +6,8 @@ from django.db import models
 from django.template import Context, Template
 from django.utils.html import strip_tags
 
-from callisto.notification.managers import EmailNotificationQuerySet
-from callisto.notification.validators import validate_email_unique
+from .managers import EmailNotificationQuerySet
+from .validators import validate_email_unique
 
 
 @six.python_2_unicode_compatible