+++ tests/callistocore/test_matching.py
@@ -8,9 +8,9 @@ from django.core.management import call_command
 from django.test import TestCase, override_settings
 from django.utils import timezone
 
-from callisto.delivery.matching import MatchingApi
-from callisto.delivery.models import MatchReport, Report
-from callisto.delivery.report_delivery import MatchReportContent
+from callisto_core.utils.api import MatchingApi
+from callisto_core.delivery.models import MatchReport, Report
+from callisto_core.delivery.report_delivery import MatchReportContent
 
 User = get_user_model()
 
@@ -36,7 +36,8 @@ class MatchTest(TestCase):
         return match_report
 
 
-@patch('callisto.delivery.matching.CallistoMatching.process_new_matches')
+@override_settings(CALLISTO_MATCHING_API='tests.callistocore.forms.CustomMatchingApi')
+@patch('tests.callistocore.forms.CustomMatchingApi.process_new_matches')
 class MatchDiscoveryTest(MatchTest):
 
     def test_running_matching_sets_report_seen(self, mock_process):
@@ -44,7 +45,7 @@ class MatchDiscoveryTest(MatchTest):
         self.create_match(self.user2, 'dummy2')
         for report in MatchReport.objects.all():
             self.assertFalse(report.seen)
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         for report in MatchReport.objects.all():
             self.assertTrue(report.seen)
 
@@ -53,14 +54,14 @@ class MatchDiscoveryTest(MatchTest):
         self.create_match(self.user2, 'dummy2')
         for report in MatchReport.objects.all():
             self.assertIsNotNone(report.identifier)
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         for report in MatchReport.objects.all():
             self.assertIsNone(report.identifier)
 
     def test_two_matching_reports_match(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_process.assert_called_once_with([match1, match2], 'dummy')
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -70,7 +71,7 @@ class MatchDiscoveryTest(MatchTest):
     def test_non_matching_reports_dont_match(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy1')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -81,7 +82,7 @@ class MatchDiscoveryTest(MatchTest):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy1')
         match3 = self.create_match(self.user2, 'dummy1')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -93,11 +94,11 @@ class MatchDiscoveryTest(MatchTest):
     def test_existing_match_not_retriggered_by_same_reporter(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertTrue(mock_process.called)
         mock_process.reset_mock()
         match3 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -109,12 +110,12 @@ class MatchDiscoveryTest(MatchTest):
     def test_triggers_new_matches_only(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertTrue(mock_process.called)
         mock_process.reset_mock()
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         self.create_match(user3, 'dummy1')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -124,13 +125,13 @@ class MatchDiscoveryTest(MatchTest):
     def test_three_way_match(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy1')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         match3 = self.create_match(user3, 'dummy1')
         user4 = User.objects.create_user(username="mmudy", password="dummy")
         match4 = self.create_match(user4, 'dummy1')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_process.assert_called_once_with([match2, match3, match4], 'dummy1')
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -146,7 +147,7 @@ class MatchDiscoveryTest(MatchTest):
         match2 = self.create_match(self.user2, 'dummy')
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         match3 = self.create_match(user3, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_process.assert_called_once_with([match1, match2, match3], 'dummy')
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -158,23 +159,23 @@ class MatchDiscoveryTest(MatchTest):
     def test_existing_match_still_triggers_on_new(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertTrue(mock_process.called)
         mock_process.reset_mock()
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         match3 = self.create_match(user3, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_process.assert_called_once_with([match1, match2, match3], 'dummy')
 
     def test_double_match(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy1')
         match2 = self.create_match(self.user2, 'dummy2')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_process.called)
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         match3 = self.create_match(user3, 'dummy1')
         match4 = self.create_match(user3, 'dummy2')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         call1 = call([match1, match3], 'dummy1')
         call2 = call([match2, match4], 'dummy2')
         mock_process.assert_has_calls([call1, call2])
@@ -184,7 +185,7 @@ class MatchDiscoveryTest(MatchTest):
         match2 = self.create_match(self.user2, 'dummy')
         mock_process.side_effect = [Exception('Boom!'), ()]
         try:
-            MatchingApi().run_matching()
+            MatchingApi.run_matching()
         except:
             pass
         match1.report.refresh_from_db()
@@ -192,7 +193,7 @@ class MatchDiscoveryTest(MatchTest):
         self.assertFalse(match1.report.match_found)
         self.assertFalse(match2.report.match_found)
         mock_process.reset_mock()
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_process.assert_called_once_with([match1, match2], 'dummy')
         match1.report.refresh_from_db()
         match2.report.refresh_from_db()
@@ -200,10 +201,9 @@ class MatchDiscoveryTest(MatchTest):
         self.assertTrue(match2.report.match_found)
 
 
-@patch('callisto.notification.api.NotificationApi.send_match_notification')
-@patch('callisto.notification.api.NotificationApi.send_email_to_authority_intake')
+@patch('callisto_core.notification.api.CallistoCoreNotificationApi.send_match_notification')
+@patch('callisto_core.notification.api.CallistoCoreNotificationApi.send_email_to_authority_intake')
 @override_settings(CALLISTO_NOTIFICATION_API='tests.callistocore.forms.SiteAwareNotificationApi')
-@override_settings(CALLISTO_MATCHING_API='callisto.delivery.matching.CallistoMatching')
 class MatchNotificationTest(MatchTest):
 
     def setUp(self):
@@ -213,7 +213,7 @@ class MatchNotificationTest(MatchTest):
     def test_both_new_matches_sent_emails(self, mock_send_to_authority, mock_send_email):
         report1 = self.create_match(self.user1, 'dummy')
         report2 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         calls = [call(self.user1, report1), call(self.user2, report2)]
         mock_send_email.assert_has_calls(calls)
         self.assertEqual(mock_send_email.call_count, 2)
@@ -223,7 +223,7 @@ class MatchNotificationTest(MatchTest):
         report2 = self.create_match(self.user2, 'dummy')
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         report3 = self.create_match(user3, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         calls = [call(self.user1, report1), call(self.user2, report2), call(user3, report3)]
         mock_send_email.assert_has_calls(calls)
         self.assertEqual(mock_send_email.call_count, 3)
@@ -231,21 +231,21 @@ class MatchNotificationTest(MatchTest):
     def test_only_new_matches_sent_emails(self, mock_send_to_authority, mock_send_email):
         self.create_match(self.user1, 'dummy')
         self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertTrue(mock_send_email.called)
         mock_send_email.reset_mock()
         user3 = User.objects.create_user(username="yumdm", password="dummy")
         report3 = self.create_match(user3, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_send_email.assert_called_once_with(user3, report3)
 
     def test_users_are_deduplicated(self, mock_send_to_authority, mock_send_email):
         report1 = self.create_match(self.user1, 'dummy')
         self.create_match(self.user1, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         self.assertFalse(mock_send_email.called)
         report3 = self.create_match(self.user2, 'dummy')
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         calls = [call(self.user1, report1), call(self.user2, report3)]
         mock_send_email.assert_has_calls(calls)
         self.assertEqual(mock_send_email.call_count, 2)
@@ -255,13 +255,13 @@ class MatchNotificationTest(MatchTest):
         report2 = self.create_match(self.user2, 'dummy')
         report2.report.submitted_to_school = timezone.now()
         report2.report.save()
-        MatchingApi().run_matching()
+        MatchingApi.run_matching()
         mock_send_email.assert_called_once_with(self.user1, report1)
 
 
 class MatchingCommandTest(MatchTest):
 
-    @patch('callisto.delivery.matching.CallistoMatching.process_new_matches')
+    @patch('callisto_core.delivery.api.CallistoCoreMatchingApi.process_new_matches')
     def test_command_runs_matches(self, mock_process):
         match1 = self.create_match(self.user1, 'dummy')
         match2 = self.create_match(self.user2, 'dummy')