+++ tests/notification/test_sites.py
@@ -6,8 +6,8 @@ from django.contrib.sites.models import Site
 from django.core.urlresolvers import reverse
 from django.test import TestCase, override_settings
 
-from callisto.delivery.models import Report
-from callisto.notification.models import EmailNotification
+from callisto_core.delivery.models import Report
+from callisto_core.notification.models import EmailNotification
 
 User = get_user_model()
 
@@ -105,7 +105,7 @@ class SiteRequestTest(TestCase):
         response = self.client.get(self.submit_url)
         self.assertNotEqual(response.status_code, 404)
 
-    @patch('callisto.notification.managers.EmailNotificationQuerySet.on_site')
+    @patch('callisto_core.notification.managers.EmailNotificationQuerySet.on_site')
     def test_site_passed_to_email_notification_manager(self, mock_on_site):
         self.client.post(
             self.submit_url,